+++ src/components/bs5/card/Card.mdx
@@ -1,5 +1,5 @@
 import { Canvas, Meta, Story } from "@storybook/blocks"
-import * as CardStories from "./card.stories"
+import * as CardStories from "./card--no-action.stories"
 
 <Meta of={CardStories} />
